update test case to combine three cases into one, reduce the time cost of inserting data by charles
